procalcs-pdf-cleaner/backend/requirements.txt
# ProCalcs PDF Cleaner — Python Dependencies

# Web framework
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.1.0
gunicorn==23.0.0

# DXF processing
ezdxf==0.19.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
